lambda/src/application/dr_dag_ntw_gw_calls/helpers/requirements.txt
snowflake-connector-python=3.2.0
jinja2=3.1.2
markupsafe=2.1.3